updated static files

+++ blog/views.py
@@ -31,7 +31,7 @@ def api_detail_blog_view(request, slug):
 
     if request.method == "GET":
         serializer = BlogPostSerializer(blog_post)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 @api_view(["PUT"])
@@ -54,18 +54,18 @@ def api_update_blog_view(request, slug):
 
         if serializer.is_valid():
             serializer.save()
-            data['success'] = UPDATE_SUCCESS
+            data['response'] = UPDATE_SUCCESS
             data['id'] = blog_post.id
             data['title'] = blog_post.title
             data['body'] = blog_post.body
             data['slug'] = blog_post.slug
-            data['updated'] = blog_post.date_updated
+            data['date_updated'] = blog_post.date_updated
             image_url = str(request.build_absolute_uri(blog_post.image.url))
             if "?" in image_url:
                 image_url = image_url[:image_url.rfind("?")]
             data['image'] = image_url
             data['username'] = blog_post.author.username
-            return Response(data=data)
+            return Response(data=data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -77,14 +77,15 @@ def api_delete_blog_view(request, slug):
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
-    if request.method == "DELETE":
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You don't have permission to delete that."})
+
+    if request.method == 'DELETE':
         operation = blog_post.delete()
         data = {}
-
         if operation:
-            data["success"] = "Post deleted!"
-        else:
-            data["failure"] = "An error occurred!"
+            data['response'] = DELETE_SUCCESS
         return Response(data=data)
 
 
@@ -123,3 +124,20 @@ class ApiBlogListView(ListAPIView):
     pagination_class = PageNumberPagination
     filter_backends = (SearchFilter, OrderingFilter)
     search_fields = ('title', 'body', 'author__username')
+
+
+@api_view(['GET', ])
+@permission_classes((IsAuthenticated,))
+def api_is_author_of_blogpost(request, slug):
+    try:
+        blog_post = BlogPost.objects.get(slug=slug)
+    except BlogPost.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    data = {}
+    user = request.user
+    if blog_post.author != user:
+        data['response'] = "You don't have permission to edit that."
+        return Response(data=data)
+    data['response'] = "You have permission to edit that."
+    return Response(data=data)
